refactor: constants.tsに分ける
